\name{gcmdiag-package}
\alias{gcmdiag-package}
\docType{package}
\title{Growth curve modeling diagnostics}
\description{Outlier and leverage observation diagnostics for growth curve modeling using six methods}
\details{
\tabular{ll}{
Package: \tab gcmdiag\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-10-08\cr
License: \tab GPL-2\cr
}
This package requires to load packages "MASS" and "diagsem". By selecting methods in the gcmdiag function, different methods can be used to identify outlying observations in linear growth curve modeling.
}
\author{
Xin Tong and Zhiyong Zhang

Maintainer: Xin Tong <xtong@virginia.edu>

}
\references{
Tong, X. and Zhang, Z. (2015). Outlying observation diagnostics in growth curve modeling.

Yuan, K.-H., & Zhang, Z. (2012). Structural equation modeling diagnostics using R package semdiag and EQS. Structural Equation Modeling: An Interdisciplinary Journal, 19(4), 683-702.
}

\keyword{Outlying observation}
\keyword{Outlier}
\keyword{Leverage observation}