\name{N300}
\alias{N300}
\docType{data}
\title{Simulated data}
\usage{
data(N300)
}

\description{
A simulated dataset which contains outlying observations.
}
\format{
  A data frame with 300 observations on the following 4 variables.
  \describe{
    \item{\code{V1}}{a numeric vector}
    \item{\code{V2}}{a numeric vector}
    \item{\code{V3}}{a numeric vector}
    \item{\code{V4}}{a numeric vector}
  }
}

